Docker and normal utilization clarified
Docker and normal utilization explained
Links to Quay provided
Postman collection referenced
Name of file and api url clarified
