+++ app.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import falcon
-from quote_resource import QuoteResource 
+from multiple_linear_resource import MultipleLinearResource 
 
 
 
 
 api = falcon.API()
-api.add_route('/quote', QuoteResource())
+api.add_route('/multiple-linear', MultipleLinearResource())
